backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
numpy==1.26.2
rembg==2.0.50
scipy==1.11.4

